+++ template-parts/pages/work-training/diversity-slider.php
@@ -59,7 +59,7 @@ if ( ! $wtd_title || ! have_rows( 'work_training_diversity_items' ) ) {
 								<?php if ( get_sub_field( 'category' ) || get_sub_field( 'title' ) ) : ?>
 									<figcaption class="section-diversity-slider__caption absolute bottom-0 left-0 w-full">
 										<?php if ( get_sub_field( 'category' ) ) : ?>
-											<span class="uppercase"><?php echo esc_html( get_sub_field( 'category' ) ); ?></span>
+											<span class="uppercase text-brand-dark"><?php echo esc_html( get_sub_field( 'category' ) ); ?></span>
 										<?php endif; ?>
 										<?php if ( get_sub_field( 'title' ) ) : ?>
 											<span><?php echo esc_html( get_sub_field( 'title' ) ); ?></span>